[SPARK-47907][SQL] Put bang under a config
### What changes were proposed in this pull request?

Recently a bug was fixed in the lexer which equated ! (bang) with NOT. The original bug allowed for unintended syntax such as:
CREATE TABLE IF ! EXISTS TY(c1 INT ! NULL)
and also, slightly less iffy: c1 ! IN (1, 2)

It turns out this undocumented "feature" can be found in the wild with examples of `! LIKE`, `! BETWEEN` and `! IN`.
So we need to provide a safety valve by allowing users to re-enable the bug, at least for a while.

### Why are the changes needed?

### Does this PR introduce _any_ user-facing change?

Yes (The original bugfix does, here we make it offical)

### How was this patch tested?

Added automated  testcases for both modes.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #46138 from srielau/SPARK-47907-config-for-bang.

Authored-by: Serge Rielau <serge@rielau.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>