[SPARK-47941][SS][CONNECT] Propagate ForeachBatch worker initialization errors to users for PySpark
### What changes were proposed in this pull request?
This change was made to propagate errors in PySpark foreachBatch worker initialization to the user, which can happen if the UDF isn't serializable or deserializable.

### Why are the changes needed?

When the foreachBatch function isn't serializable, previously we were not propagating this error to the user, and instead were failing with an ambiguous SparkConnectGrpcException (while the real error was being written to stderr). With this change, we will send the error message to the user with a more specific error message about worker initialization failure.

### Does this PR introduce _any_ user-facing change?

Yes, we are simply propagating the error to the user instead of just writing the error to stderr.

### How was this patch tested?

Using unit tests and a notebook with the change.
Stack trace before change:
```
SparkConnectGrpcException: (java.io.EOFException)
File <command-41186175708098>, line 27
      def fcn(batch, id):
          print(obj)
       q = (df.select("value")
          .writeStream
          .foreachBatch(fcn)
--->       .start())
```

Stack trace after change:
```
SparkConnectGrpcException: (java.lang.RuntimeException) Runner initialization failed (returned -2). Error message: Traceback (most recent call last):
  File "/databricks/spark/python/pyspark/serializers.py", line 192, in _read_with_length
    return self.loads(obj)
           ^^^^^^^^^^^^^^^
  File "/databricks/spark/python/pyspark/serializers.py", line 572, in loads
    return cloudpickle.loads(obj, encoding=encoding)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/spark-f94b69d4-fdb2-409f-89cf-1a/.ipykernel/13276/command-41186175708098-95633136", line 4, in _reduce_function
ValueError: Cannot unpickle this object
```
### Was this patch authored or co-authored using generative AI tooling?

No

Closes #46125 from ericm-db/feb-error-prop.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>