[SPARK-47963][CORE] Make the external Spark ecosystem can use structured logging mechanisms
### What changes were proposed in this pull request?
The pr aims to make `the external Spark ecosystem` can use `structured logging mechanisms`.

### Why are the changes needed?
Currently, the class of MDC's key is `LogKey`, while LogKey is an `enumeration`. If we want to use structured logging mechanism in a `third-party Spark ecosystem`, we cannot currently use it (we cannot expand and `add a new LogKey enumeration value`)
So I propose changing the `LogKey` from `enumeration` to the following,
<img width="550" alt="image" src="https://github.com/apache/spark/assets/15246973/ffb0f41e-4f2c-4537-b43f-d55353fe03f7">
as we already have a similar way of writing it in the Spark code, eg:
https://github.com/apache/spark/blob/3f088d5ff9737a130e825bc466109236f27d1858/core/src/main/scala/org/apache/spark/TaskEndReason.scala#L39-L46

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
- Add new UT (Give an example to developers who use the LogKey mechanism in external systems related to Spark.)
- Manually test.
- Pass GA.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #46193 from panbingkun/SPARK-47963.

Lead-authored-by: panbingkun <panbingkun@baidu.com>
Co-authored-by: panbingkun <pbk1982@gmail.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>