[SPARK-47994][SQL] Fix bug with CASE WHEN column filter push down in SQLServer
### What changes were proposed in this pull request?

In this PR I propose a change in QueryBuilder for SQLServer. This change modifies push down of predicate that contains a column that is generated with a CASE WHEN construct, so that we add simple ` = 1` comparison to this query, making it work on SQLServer.

### Why are the changes needed?

SQLServer does not support 0 or 1 as a boolean values. There are certain situations where spark optimizer rewrites filters that contain CASE WHEN columns in a way that adds 1 or 0 as a boolean values, which fails on SQLServer side with an error "An expression of non-boolean type specified in a context where a condition is expected". With these changes, we modify pushing this filters down, and error is no longer present.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?
A new test case is added, which fails without these changes.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46231 from stefanbuk-db/SQLServer_case_when_bugfix.

Authored-by: Stefan Bukorovic <stefan.bukorovic@databricks.com>
Signed-off-by: Kent Yao <yao@apache.org>