[SPARK-48067][SQL] Fix variant default columns
### What changes were proposed in this pull request?

Changes the literal `sql` representation of a variant value to `parse_json(variant.toJson)`. This is because there is no other representation of a literal variant.

This allows variant default columns to work because default columns store a literal string representation in the schema struct fields metadata as the default value.

### Why are the changes needed?

previously we could not set a variant default column like
```
create table t(
        v6 variant default parse_json('{\"k\": \"v\"}')
)
```

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

added UT

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #46312 from richardc-db/fix_variant_default_cols.

Authored-by: Richard Chen <r.chen@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>