[SPARK-48059][CORE] Implement the structured log framework on the java side
### What changes were proposed in this pull request?
The pr aims to implement the structured log framework on the `java side`.

### Why are the changes needed?
Currently, the structured log framework on the `scala side` is basically available, but the`Spark Core` code also includes some `Java code`, which also needs to be connected to the structured log framework.

### Does this PR introduce _any_ user-facing change?
Yes, only for developers.

### How was this patch tested?
- Add some new UT.
- Pass GA.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #46301 from panbingkun/structured_logger_java.

Authored-by: panbingkun <panbingkun@baidu.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>