[SPARK-48192][INFRA] Enable TPC-DS tests in forked repository
### What changes were proposed in this pull request?

This PR is a sort of a followup of https://github.com/apache/spark/pull/46361. It proposes to run TPC-DS and Docker integration tests in PRs (that does not consume ASF resources).

### Why are the changes needed?

TPC-DS and Docker integration stuff at least have to be tested in the PR if the PR touches the codes related to that.

### Does this PR introduce _any_ user-facing change?

No, test-only.

### How was this patch tested?

Manually

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46470 from HyukjinKwon/SPARK-48192.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>