[SPARK-48341][CONNECT] Allow plugins to use QueryTest in their tests
### What changes were proposed in this pull request?

This PR changes `QueryTest` to no longer depend on `RemoteSparkSession`.

### Why are the changes needed?

This allows the tests for Spark Connect plugin to provide their version of `RemoteSparkSession` (which depends on some idiosyncrasies of how Spark is built).

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests to ensure that nothing breaks. Manually tested that this allows a plugin to use `QueryTest`.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #46667 from tomvanbussel/SPARK-48341.

Lead-authored-by: Tom van Bussel <tom.vanbussel@databricks.com>
Co-authored-by: Tom van Bussel <Tom.VanBussel@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>