[SPARK-48379][INFRA] Cancel previous builds on a PR when new commit is pushed
### What changes were proposed in this pull request?

When there is an open PR with a build in progress and a new commit is pushed I propose that we should cancel the previous build.

The exceptions are branches `master` and `branch-*` where we still want to be able to have multiple builds executing concurrently (if two PRs are merged to main at the same time we don't want to just run the build on the second one).

[Concurrecy docs](https://docs.github.com/en/actions/using-jobs/using-concurrency)

### Why are the changes needed?

To reduce wait times for newer commit's and save compute resources.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Manually.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46689 from stefankandic/cancelOlderBuilds.

Lead-authored-by: Stefan Kandic <stefan.kandic@databricks.com>
Co-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>