[MINOR][INFRA] Make pure Python build compatible with other setuptools versions
### What changes were proposed in this pull request?

This PR is related to https://github.com/apache/spark/pull/46751. For now, we're using low `setuptools` version so the build is not affected, but it will be broken if we happen to upgrade it at some point. This PR makes the build compatible with `_` and `-`.

### Why are the changes needed?

This is a proactive action to prevent breaking the build.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Manually

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46752 from HyukjinKwon/minor-build.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>