[SPARK-48442][PYTHON][TESTS] Add parenthesis to awaitTermination
### Why are the changes needed?
In `test_stream_reader` and `test_stream_writer` of `test_python_streaming_datasource.py`, the call `q.awaitTermination` does not invoke a function call as intended, but instead returns a python function object. The fix is to change this to `q.awaitTermination()`.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Tests were not added since it is correcting a bug in a test file.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46779 from riyaverm-db/bug-fix-add-paren-to-awaitTermination.

Authored-by: Riya Verma <riya.verma@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>