[SPARK-48446][SS][DOCS] Update SS doc of dropDuplicates to use the right syntax
### What changes were proposed in this pull request?
This PR fixes the wrong usage of `dropDuplicates` and `dropDuplicatesWithinWatermark` in the Structured Streaming Programming Guide.

### Why are the changes needed?
Previously the syntax in the guide was wrong, so users will see an error if directly using the example.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Made sure that the updated examples conform to the API doc, and can run out of the box.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #46797 from eason-yuchen-liu/dropduplicate-doc.

Authored-by: Yuchen Liu <yuchen.liu@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>