[SPARK-47716][SQL] Avoid view name conflict in SQLQueryTestSuite semantic sort test case
### What changes were proposed in this pull request?
In SQLQueryTestSuite, the test case "Test logic for determining whether a query is semantically sorted" can sometimes failure with an error
```
Cannot create table or view `main`.`default`.`t1` because it already exists.
```
if run concurrently with other sql test cases that also create tables with the same name.

Fix it by putting it in a schema with a unique name.

### Why are the changes needed?
Fix flaky test issue

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Test itself

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #45855 from jchen5/sql-sort-test.

Authored-by: Jack Chen <jack.chen@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>