[SPARK-48392][CORE][FOLLOWUP] Add `--load-spark-defaults` flag to decide whether to load `spark-defaults.conf`
### What changes were proposed in this pull request?

Following the discussions in #46709, this PR adds a flag `--load-spark-defaults` to control whether `spark-defaults.conf` should be loaded when `--properties-file` is specified. By default, the flag is turned off.

### Why are the changes needed?

Ideally we should avoid behavior change and reduce user disruptions. For this reason, a flag is preferred.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

N/A

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #46782 from sunchao/SPARK-48392-followup.

Authored-by: Chao Sun <chao@openai.com>
Signed-off-by: Chao Sun <chao@openai.com>