[SPARK-48374][PYTHON] Support additional PyArrow Table column types
### What changes were proposed in this pull request?
This is a small follow-up to #46529. This adds support for some more Arrow data types:
- fixed-size binary
- fixed-size list
- large list

### Why are the changes needed?
Users who are creating Spark DataFrames from PyArrow Tables will expect it to work if their Tables contain these types of columns

### Does this PR introduce _any_ user-facing change?
It will prevent an error in the case where the user has one of these types of columns in their PyArrow Table. There are no other user-facing changes.

### How was this patch tested?
Tests are included.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46688 from ianmcook/SPARK-48374.

Authored-by: Ian Cook <ianmcook@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>