[SPARK-48489][SQL] Throw a better, user-facing error when reading illegal schema from text Data Source
### What changes were proposed in this pull request?
In this PR, I propose a change which would delete internal assert, and throw a user-facing error with better description if someone tries to read from `text` DataSource into table with schema of multiple columns.

### Why are the changes needed?
This changes are needed for better error handling and smoother user experience.

### Does this PR introduce _any_ user-facing change?
This PR gives user a better error message in some scenarios.

### How was this patch tested?

### Was this patch authored or co-authored using generative AI tooling?

Closes #46823 from stefanbuk-db/TextDSUserFacingError.

Authored-by: Stefan Bukorovic <stefan.bukorovic@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>